+++ lib/RankingStrategy/ModifiedCompetition.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright 2012 Maarten Claes 
+ * Copyright 2012 Maarten Claes
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- 
+
+namespace Ranker\RankingStrategy;
+
 /**
  * Modified competition strategy ( 1334 )
- * Leaves the gaps in the ranking numbers before the sets of equal-ranking 
+ * Leaves the gaps in the ranking numbers before the sets of equal-ranking
  * items (rather than after them as in standard competition ranking)
  */
-class ModifiedCompetitionStrategy extends RankingStrategy {
+class ModifiedCompetition extends Base {
 
   private $equalRanking = array();
 
   protected function whenEqual($rankable, $ranking_index) {
     $this->updateEqualRankingItems($rankable, $ranking_index);
   }
-  
+
   private function updateEqualRankingItems($rankable, $ranking_index) {
     $this->equalRanking[] = $rankable;
     // ranking = value after gap = $ranking_index + 1
